# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: MindTab API
  description: API for MindTab — task tracking and micro-notes platform.
  version: 0.1.0
  contact:
    name: MindTab
    url: https://mindtab.in
servers:
  - url: /
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /users/{id}:
    get:
      operationId: getUserById
      tags: [Users]
      summary: Get public user profile
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            maxLength: 255
      responses:
        "200":
          description: Public user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Tasks ──────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Parameters ───────────────────────────────────────────────────────
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
          description: Machine-readable error code

    # ── Auth ───────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, email, xp, onboardingCompleted, theme, font, createdAt]
      properties:
        id:
          type: string
          maxLength: 255
          description: User ID (varchar, not UUID)
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        image:
          type: string
          nullable: true
        xp:
          type: integer
          default: 0
        onboardingCompleted:
          type: boolean
          default: false
        theme:
          type: string
          enum: [midnight, graphite, paper]
          default: midnight
        font:
          type: string
          enum: [inter, geist, system]
          default: inter
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true

    # ── Task ───────────────────────────────────────────────────────────
...